+++ engines/classic-rbdoom-3-bfg/build.sh
@@ -3,7 +3,22 @@
 # CLONE PHASE
 git clone https://github.com/MadDeCoDeR/Classic-RBDOOM-3-BFG.git source
 pushd source
-git checkout 54a7d8d
+git checkout cf9c2fa
+popd
+
+git clone https://github.com/kcat/openal-soft.git openal
+pushd openal
+git checkout ae4eacf
+popd
+
+# BUILD PHASE
+pushd "openal"
+cd build
+cmake \
+    -DCMAKE_BUILD_TYPE=MinSizeRel \
+    ..
+make -j "$(nproc)"
+make install
 popd
 
 pushd "source"
@@ -23,3 +38,5 @@ popd
 cp "source/build/DoomBFA" "$diststart/208200/dist/DoomBFA"
 cp -rfv ./assets/* "$diststart/208200/dist/"
 cp -rfv ./source/base "$diststart/208200/dist/updatedbase"
+cp -rfv "/usr/local/lib/libopenal.so.1.21.1" "$diststart/208200/dist/lib/libopenal.so"
+cp -rfv "/usr/local/lib/libopenal.so.1.21.1" "$diststart/208200/dist/lib/libopenal.so.1"