+++ engines/dosbox-staging/env.sh
@@ -3,3 +3,4 @@
 export STEAM_APP_ID_LIST="default 410970"
 export LICENSE_PATH="./source/COPYING"
 export COMMON_PACKAGE="1"
+export ADDITIONAL_LICENSES="./source/subprojects/fluidsynth-2.2.0/LICENSE ./source/subprojects/munt-libmt32emu_2_5_0/LICENSE.build"