+++ engines/easyrpg-player/build.sh
@@ -3,17 +3,34 @@
 # CLONE PHASE
 git clone https://github.com/EasyRPG/Player.git source
 pushd source
-git checkout -f 4dd00a6
+git checkout -f 281be71
 git submodule update --init --recursive
 popd
 
 git clone https://github.com/EasyRPG/liblcf.git liblcf
 pushd liblcf
-git checkout -f bb9f9e2
+git checkout -f 01b73de
+popd
+
+git clone https://github.com/fmtlib/fmt.git fmt
+pushd fmt
+git checkout -f d141cdb
 popd
 
 # BUILD PHASE
 pushd liblcf
+mkdir -p build
+cd build
+cmake \
+    -DCMAKE_PREFIX_PATH="$pfx;$pfx/usr/local" \
+    -DCMAKE_INSTALL_PREFIX="$pfx" \
+    -DCMAKE_BUILD_TYPE=Release \
+    ..
+make -j "$(nproc)"
+make install
+popd
+
+pushd "fmt"
 mkdir -p build
 cd build
 cmake \