+++ engines/rbdoom-3-bfg/env.sh
@@ -2,5 +2,5 @@
 
 export STEAM_APP_ID_LIST="208200"
 export LICENSE_PATH="./source/LICENSE.md"
-export ADDITIONAL_LICENSES="./source/docs/CODE_LICENSE_EXCEPTIONS.md"
+export ADDITIONAL_LICENSES="./source/LICENSE_EXCEPTIONS.md"
 export LIBRARIES="yasm ffmpeg"