+++ engines/realrtcw/assets/setup-realrtcw.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+if [ -f "last_error.txt" ]; then
+    rm last_error.txt
+fi
+
 if [[ ! -z "${DEPPATH_9010}" ]]; then
     echo "Automatic path for rtcw found at $DEPPATH_9010"
     RTCW_PATH="$DEPPATH_9010"
@@ -9,13 +13,23 @@ else
 fi
 
 if [ -z "$RTCW_PATH" ]; then
-    "$STEAM_ZENITY" --error --title="RealRTCW Setup Error" --text="Path to RTCW not given"
-    exit 1
+    error_message="Path to RTCW not given."
+    if [[ -z "${LUX_ERRORS_SUPPORTED}" ]]; then
+        "$STEAM_ZENITY" --error --title="Error" --text="$error_message"
+    else
+        echo "$error_message" > last_error.txt
+    fi
+    exit 10
 fi
 
 if [ ! -d ./nosteam/"!copy the content of this folder into rtcw root directory" ]; then
-    "$STEAM_ZENITY" --error --title="RealRTCW Setup Error" --text="Game content not downloaded from ModDB and extracted to nosteam directory."
-    exit 1
+    error_message="Game content not downloaded from ModDB and extracted to nosteam directory."
+    if [[ -z "${LUX_ERRORS_SUPPORTED}" ]]; then
+        "$STEAM_ZENITY" --error --title="Error" --text="$error_message"
+    else
+        echo "$error_message" > last_error.txt
+    fi
+    exit 10
 fi
 
 mv ./nosteam/"!copy the content of this folder into rtcw root directory"/Main/* ./nosteam/Main